CXP-1182: Creation of the ServiceAPI
