+++ components/catalogs/back/src/Infrastructure/Symfony/Resources/config/services.yml
@@ -12,6 +12,7 @@ services:
         exclude:
             - '../../../Symfony/'
             - '../../../../Domain/'
+            - '../../../../ServiceAPI/'
             - '../../../../Application/Exception'
 
     Akeneo\Catalogs\Application\Handler\:
