+++ components/catalogs/back/src/ServiceAPI/Command/CreateCatalogCommand.php
@@ -1,0 +1,40 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Akeneo\Catalogs\ServiceAPI\Command;
+
+use Symfony\Component\Validator\Constraints as Assert;
+
+/**
+ * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+final class CreateCatalogCommand
+{
+    public function __construct(
+        #[Assert\NotBlank]
+        #[Assert\Uuid]
+        private string $id,
+        #[Assert\NotBlank]
+        #[Assert\Length(min: 1, max: 255)]
+        private string $name,
+        private int $ownerId,
+    ) {
+    }
+
+    public function getId(): string
+    {
+        return $this->id;
+    }
+
+    public function getName(): string
+    {
+        return $this->name;
+    }
+
+    public function getOwnerId(): int
+    {
+        return $this->ownerId;
+    }
+}
