+++ components/catalogs/back/src/ServiceAPI/Exception/CatalogNotFoundException.php
@@ -2,24 +2,18 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Domain\Query;
-
-use Akeneo\Catalogs\Domain\Model\Catalog;
-use Symfony\Component\Validator\Constraints as Assert;
+namespace Akeneo\Catalogs\ServiceAPI\Exception;
 
 /**
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- *
- * @implements Query<Catalog>
  */
-final class GetCatalogQuery implements Query
+class CatalogNotFoundException extends \Exception
 {
     public function __construct(
-        #[Assert\NotBlank]
-        #[Assert\Uuid]
-        private string $id,
+        private string $id
     ) {
+        parent::__construct();
     }
 
     public function getId(): string
