+++ components/catalogs/back/src/ServiceAPI/Query/GetCatalogQuery.php
@@ -2,39 +2,28 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Domain\Command;
+namespace Akeneo\Catalogs\ServiceAPI\Query;
 
+use Akeneo\Catalogs\ServiceAPI\Model\Catalog;
 use Symfony\Component\Validator\Constraints as Assert;
 
 /**
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ *
+ * @implements Query<Catalog>
  */
-final class CreateCatalogCommand
+final class GetCatalogQuery implements Query
 {
     public function __construct(
         #[Assert\NotBlank]
         #[Assert\Uuid]
         private string $id,
-        #[Assert\NotBlank]
-        #[Assert\Length(min: 1, max: 255)]
-        private string $name,
-        private int $ownerId,
     ) {
     }
 
     public function getId(): string
     {
         return $this->id;
-    }
-
-    public function getName(): string
-    {
-        return $this->name;
-    }
-
-    public function getOwnerId(): int
-    {
-        return $this->ownerId;
     }
 }
