+++ components/catalogs/back/tests/Integration/Application/Handler/CreateCatalogHanderTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Catalogs\Test\Integration\Application\Handler;
 
-use Akeneo\Catalogs\Domain\Command\CreateCatalogCommand;
+use Akeneo\Catalogs\ServiceAPI\Command\CreateCatalogCommand;
 use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
