+++ components/catalogs/back/tests/Integration/Application/Handler/GetCatalogHanderTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Catalogs\Test\Integration\Application\Handler;
 
-use Akeneo\Catalogs\Domain\Query\GetCatalogQuery;
+use Akeneo\Catalogs\ServiceAPI\Query\GetCatalogQuery;
 use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
